backend/requirements.txt
fastapi==0.95.1
uvicorn==0.20.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
psycopg2-binary==2.9.6
alembic==1.10.2
python-dotenv==1.0.0
pytest==7.2.0
python-jose
bcrypt==4.0.1
passlib[bcrypt]
python-multipart==0.0.20
pydantic[email]